Completed 'take_loan', altered 'search_book'
Finished the 'take_loan' method by using an altered version of the 'search_book' method that returns the index of the book in book_list if it exists, or 0 if it does not. This allows take_loan to quickly determine whether a book is available and print accordingly before it adds a book to the loan_list.